# docs/api.yaml
openapi: 3.1.0
info:
  title: WeFixIt HTTP API
  version: 1.0.0
  description: Session-authenticated endpoints used by the server-rendered application.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wefixit_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
paths:
...
  /support/disputes:
    post:
      summary: Open a dispute for an eligible customer booking
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '201': { description: Dispute opened }
  /support/refunds:
    post:
      summary: Request a refund against a paid booking
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '201': { description: Refund requested }
        '409': { description: Refund is ineligible or already active }
  /support/reschedules:
    post:
      summary: Propose a new appointment window
      security: [{ sessionCookie: [], csrfHeader: [] }]
      responses:
        '201': { description: Reschedule requested }
...